refac(tricky-pipe): move MPSC to `mpsc` module
Now that we have multiple modules in `tricky-pipe` that provide non-MPSC
APIs, like `serbox` and `oneshot`, it seems nicer to move the MPSC into
a `mpsc` module.
