+++ source/tricky-pipe/src/mpsc/channel_core.rs
