+++ source/tricky-pipe/src/mpsc/error.rs
