+++ source/tricky-pipe/src/mpsc/static_impl.rs
